move script executor for accounts, scripts, tx validator to backend
