+++ engine/access/rpc/backend/script_executor.go
@@ -1,4 +1,4 @@
-package scripts
+package backend
 
 import (
 	"context"
