+++ engine/access/rpc/backend/script_executor_test.go
@@ -1,4 +1,4 @@
-package scripts
+package backend
 
 import (
 	"context"
